+++ pcs/lib/commands/stonith.py
@@ -1,9 +1,15 @@
-from typing import Container, Iterable, Optional
+from collections import Counter
+from typing import Container, Iterable, List, Optional, Set, Tuple
+
+from lxml.etree import _Element
 
 from pcs.common import reports
+from pcs.common.reports import ReportItemList
+from pcs.common.reports import ReportProcessor
 from pcs.common.reports.item import ReportItem
 from pcs.lib.cib import resource
 from pcs.lib.cib import stonith
+from pcs.lib.cib.nvpair import INSTANCE_ATTRIBUTES_TAG, get_value
 from pcs.lib.cib.resource.common import are_meta_disabled
 from pcs.lib.cib.tools import IdProvider
 from pcs.lib.commands.resource import (
@@ -20,6 +26,7 @@ from pcs.lib.communication.tools import (
 )
 from pcs.lib.env import LibraryEnvironment
 from pcs.lib.errors import LibraryError
+from pcs.lib.external import CommandRunner
 from pcs.lib.node import get_existing_nodes_names
 from pcs.lib.pacemaker.live import (
     FenceHistoryCommandErrorException,
@@ -268,55 +275,195 @@ def history_update(env: LibraryEnvironment):
         ) from e
 
 
-def update_scsi_devices(
-    env: LibraryEnvironment,
-    stonith_id: str,
-    set_device_list: Iterable[str],
-    force_flags: Container[reports.types.ForceCode] = (),
-) -> None:
+def _validate_add_remove_items(
+    add_item_list: Iterable[str],
+    remove_item_list: Iterable[str],
+    current_item_list: Iterable[str],
+    container_type: reports.types.AddRemoveContainerType,
+    item_type: reports.types.AddRemoveItemType,
+    container_id: str,
+    adjacent_item_id: Optional[str] = None,
+    container_can_be_empty: bool = False,
+) -> ReportItemList:
     """
-    Update scsi fencing devices without restart and affecting other resources.
+    Validate if items can be added or removed to or from a container.
 
-    env -- provides all for communication with externals
-    stonith_id -- id of stonith resource
-    set_device_list -- paths to the scsi devices that would be set for stonith
-        resource
-    force_flags -- list of flags codes
+    add_item_list -- items to be added
+    remove_item_list -- items to be removed
+    current_item_list -- items currently in the container
+    container_type -- container type
+    item_type -- item type
+    container_id -- id of the container
+    adjacent_item_id -- an adjacent item in the container
+    container_can_be_empty -- flag to decide if container can be left empty
     """
-    if not is_getting_resource_digest_supported(env.cmd_runner()):
+    # pylint: disable=too-many-locals
+    report_list: ReportItemList = []
+    if not add_item_list and not remove_item_list:
+        report_list.append(
+            ReportItem.error(
+                reports.messages.AddRemoveItemsNotSpecified(
+                    container_type, item_type, container_id
+                )
+            )
+        )
+
+    def _get_duplicate_items(item_list: Iterable[str]) -> Set[str]:
+        return {item for item, count in Counter(item_list).items() if count > 1}
+
+    duplicate_items_list = _get_duplicate_items(
+        add_item_list
+    ) | _get_duplicate_items(remove_item_list)
+    if duplicate_items_list:
+        report_list.append(
+            ReportItem.error(
+                reports.messages.AddRemoveItemsDuplication(
+                    container_type,
+                    item_type,
+                    container_id,
+                    sorted(duplicate_items_list),
+                )
+            )
+        )
+    already_present = set(add_item_list).intersection(current_item_list)
+    # report only if an adjacent id is not defined, because we want to allow
+    # to move items when adjacent_item_id is specified
+    if adjacent_item_id is None and already_present:
+        report_list.append(
+            ReportItem.error(
+                reports.messages.AddRemoveCannotAddItemsAlreadyInTheContainer(
+                    container_type,
+                    item_type,
+                    container_id,
+                    sorted(already_present),
+                )
+            )
+        )
+    missing_items = set(remove_item_list).difference(current_item_list)
+    if missing_items:
+        report_list.append(
+            ReportItem.error(
+                reports.messages.AddRemoveCannotRemoveItemsNotInTheContainer(
+                    container_type,
+                    item_type,
+                    container_id,
+                    sorted(missing_items),
+                )
+            )
+        )
+    common_items = set(add_item_list) & set(remove_item_list)
+    if common_items:
+        report_list.append(
+            ReportItem.error(
+                reports.messages.AddRemoveCannotAddAndRemoveItemsAtTheSameTime(
+                    container_type,
+                    item_type,
+                    container_id,
+                    sorted(common_items),
+                )
+            )
+        )
+    if not container_can_be_empty and not add_item_list:
+        remaining_items = set(current_item_list).difference(remove_item_list)
+        if not remaining_items:
+            report_list.append(
+                ReportItem.error(
+                    reports.messages.AddRemoveCannotRemoveAllItemsFromTheContainer(
+                        container_type,
+                        item_type,
+                        container_id,
+                        list(current_item_list),
+                    )
+                )
+            )
+    if adjacent_item_id:
+        if adjacent_item_id not in current_item_list:
+            report_list.append(
+                ReportItem.error(
+                    reports.messages.AddRemoveAdjacentItemNotInTheContainer(
+                        container_type,
+                        item_type,
+                        container_id,
+                        adjacent_item_id,
+                    )
+                )
+            )
+        if adjacent_item_id in add_item_list:
+            report_list.append(
+                ReportItem.error(
+                    reports.messages.AddRemoveCannotPutItemNextToItself(
+                        container_type,
+                        item_type,
+                        container_id,
+                        adjacent_item_id,
+                    )
+                )
+            )
+        if not add_item_list:
+            report_list.append(
+                ReportItem.error(
+                    reports.messages.AddRemoveCannotSpecifyAdjacentItemWithoutItemsToAdd(
+                        container_type,
+                        item_type,
+                        container_id,
+                        adjacent_item_id,
+                    )
+                )
+            )
+    return report_list
+
+
+def _update_scsi_devices_get_element_and_devices(
+    runner: CommandRunner,
+    report_processor: ReportProcessor,
+    cib: _Element,
+    stonith_id: str,
+) -> Tuple[_Element, List[str]]:
+    """
+    Do checks and return stonith element and list of current scsi devices.
+    Raise LibraryError if checks fail.
+
+    runner -- command runner instance
+    report_processor -- tool for warning/info/error reporting
+    cib -- cib element
+    stonith_id -- id of stonith resource
+    """
+    if not is_getting_resource_digest_supported(runner):
         raise LibraryError(
             ReportItem.error(
                 reports.messages.StonithRestartlessUpdateOfScsiDevicesNotSupported()
-            )
-        )
-    cib = env.get_cib()
-    if not set_device_list:
-        env.report_processor.report(
-            ReportItem.error(
-                reports.messages.InvalidOptionValue(
-                    "devices", "", None, cannot_be_empty=True
-                )
             )
         )
     (
         stonith_el,
         report_list,
     ) = stonith.validate_stonith_restartless_update(cib, stonith_id)
-    if env.report_processor.report_list(report_list).has_errors:
+    if report_processor.report_list(report_list).has_errors:
         raise LibraryError()
-    # for mypy, this should not happen because exeption would be raised
+    # for mypy, this should not happen because exception would be raised
     if stonith_el is None:
         raise AssertionError("stonith element is None")
-
-    stonith.update_scsi_devices_without_restart(
-        env.cmd_runner(),
-        env.get_cluster_state(),
-        stonith_el,
-        IdProvider(cib),
-        set_device_list,
+    current_device_list = get_value(
+        INSTANCE_ATTRIBUTES_TAG, stonith_el, "devices"
     )
+    if current_device_list is None:
+        raise AssertionError("current_device_list is None")
+    return stonith_el, current_device_list.split(",")
 
-    # Unfencing
+
+def _unfencing_scsi_devices(
+    env: LibraryEnvironment,
+    device_list: Iterable[str],
+    force_flags: Container[reports.types.ForceCode] = (),
+) -> None:
+    """
+    Unfence scsi devices provided in device_list if it is possible to connect
+    to pcsd and corosync is running.
+
+    env -- provides all for communication with externals
+    device_list -- devices to be unfenced
+    force_flags -- list of flags codes
+    """
     cluster_nodes_names, nodes_report_list = get_existing_nodes_names(
         env.get_corosync_conf(),
         error_on_missing_name=True,
@@ -340,8 +487,104 @@ def update_scsi_devices(
     online_corosync_target_list = run_and_raise(
         env.get_node_communicator(), com_cmd
     )
-    com_cmd = Unfence(env.report_processor, sorted(set_device_list))
+    com_cmd = Unfence(env.report_processor, sorted(device_list))
     com_cmd.set_targets(online_corosync_target_list)
     run_and_raise(env.get_node_communicator(), com_cmd)
 
+
+def update_scsi_devices(
+    env: LibraryEnvironment,
+    stonith_id: str,
+    set_device_list: Iterable[str],
+    force_flags: Container[reports.types.ForceCode] = (),
+) -> None:
+    """
+    Update scsi fencing devices without restart and affecting other resources.
+
+    env -- provides all for communication with externals
+    stonith_id -- id of stonith resource
+    set_device_list -- paths to the scsi devices that would be set for stonith
+        resource
+    force_flags -- list of flags codes
+    """
+    if not set_device_list:
+        env.report_processor.report(
+            ReportItem.error(
+                reports.messages.InvalidOptionValue(
+                    "devices", "", None, cannot_be_empty=True
+                )
+            )
+        )
+    runner = env.cmd_runner()
+    (
+        stonith_el,
+        current_device_list,
+    ) = _update_scsi_devices_get_element_and_devices(
+        runner, env.report_processor, env.get_cib(), stonith_id
+    )
+    if env.report_processor.has_errors:
+        raise LibraryError()
+    stonith.update_scsi_devices_without_restart(
+        runner,
+        env.get_cluster_state(),
+        stonith_el,
+        IdProvider(stonith_el),
+        set_device_list,
+    )
+    devices_for_unfencing = set(set_device_list).difference(current_device_list)
+    if devices_for_unfencing:
+        _unfencing_scsi_devices(env, devices_for_unfencing, force_flags)
+    env.push_cib()
+
+
+def update_scsi_devices_add_remove(
+    env: LibraryEnvironment,
+    stonith_id: str,
+    add_device_list: Iterable[str],
+    remove_device_list: Iterable[str],
+    force_flags: Container[reports.types.ForceCode] = (),
+) -> None:
+    """
+    Update scsi fencing devices without restart and affecting other resources.
+
+    env -- provides all for communication with externals
+    stonith_id -- id of stonith resource
+    add_device_list -- paths to the scsi devices that would be added to the
+        stonith resource
+    remove_device_list -- paths to the scsi devices that would be removed from
+        the stonith resource
+    force_flags -- list of flags codes
+    """
+    runner = env.cmd_runner()
+    (
+        stonith_el,
+        current_device_list,
+    ) = _update_scsi_devices_get_element_and_devices(
+        runner, env.report_processor, env.get_cib(), stonith_id
+    )
+    if env.report_processor.report_list(
+        _validate_add_remove_items(
+            add_device_list,
+            remove_device_list,
+            current_device_list,
+            reports.const.ADD_REMOVE_CONTAINER_TYPE_STONITH_RESOURCE,
+            reports.const.ADD_REMOVE_ITEM_TYPE_DEVICE,
+            stonith_el.get("id", ""),
+        )
+    ).has_errors:
+        raise LibraryError()
+    updated_device_set = (
+        set(current_device_list)
+        .union(add_device_list)
+        .difference(remove_device_list)
+    )
+    stonith.update_scsi_devices_without_restart(
+        env.cmd_runner(),
+        env.get_cluster_state(),
+        stonith_el,
+        IdProvider(stonith_el),
+        updated_device_set,
+    )
+    if add_device_list:
+        _unfencing_scsi_devices(env, add_device_list, force_flags)
     env.push_cib()